requirements.txt
wandb~=0.18.3
lightning~=2.4.0
pillow~=10.4.0
mmgen~=0.7.3
numpy~=2.0.1
omegaconf~=2.3.0
pydantic~=2.9.2
plyfile~=1.1
torchvision~=0.19.1
tqdm~=4.65.2
imageio~=2.35.1
opencv-python~=4.10.0.84
requests~=2.28.2
scipy~=1.14.1
prettytable~=3.11.0
click~=8.1.7
mmcls~=0.25.0
pytest~=8.3.3
setuptools~=60.2.0